Store time as seconds as a long in map store
This avoids overflowing the value in 2038.

Closes #10960

+++ model/map-hot-rod/src/main/java/org/keycloak/models/map/storage/hotRod/user/HotRodUserEntity.java
@@ -153,7 +153,7 @@ public class HotRodUserEntity extends AbstractHotRodEntity {
     public String serviceAccountClientLink;
 
     @ProtoField(number = 21)
-    public Integer notBefore;
+    public Long notBefore;
 
     public static abstract class AbstractHotRodUserEntityDelegate extends UpdatableHotRodEntityDelegateImpl<HotRodUserEntity> implements MapUserEntity {
