+++ model/map-jpa/src/main/java/org/keycloak/models/map/storage/jpa/authSession/entity/JpaRootAuthenticationSessionEntity.java
@@ -78,7 +78,7 @@ public class JpaRootAuthenticationSessionEntity extends AbstractRootAuthenticati
 
     @Column(insertable = false, updatable = false)
     @Basic(fetch = FetchType.LAZY)
-    private Integer timestamp;
+    private Long timestamp;
 
     @Column(insertable = false, updatable = false)
     @Basic(fetch = FetchType.LAZY)
@@ -102,7 +102,7 @@ public class JpaRootAuthenticationSessionEntity extends AbstractRootAuthenticati
      * Used by hibernate when calling cb.construct from read(QueryParameters) method.
      * It is used to select root auth session without metadata(json) field.
      */
-    public JpaRootAuthenticationSessionEntity(UUID id, Integer entityVersion, String realmId, Integer timestamp, Long expiration) {
+    public JpaRootAuthenticationSessionEntity(UUID id, Integer entityVersion, String realmId, Long timestamp, Long expiration) {
         this.id = id;
         this.entityVersion = entityVersion;
         this.realmId = realmId;
@@ -158,13 +158,13 @@ public class JpaRootAuthenticationSessionEntity extends AbstractRootAuthenticati
     }
 
     @Override
-    public Integer getTimestamp() {
+    public Long getTimestamp() {
         if (isMetadataInitialized()) return metadata.getTimestamp();
         return timestamp;
     }
 
     @Override
-    public void setTimestamp(Integer timestamp) {
+    public void setTimestamp(Long timestamp) {
         metadata.setTimestamp(timestamp);
     }
